added a test case for an api to receive money.

+++ api/src/main/java/com/base/controller/MoneySharingController.java
@@ -13,7 +13,7 @@ import reactor.core.publisher.Mono;
 public class MoneySharingController {
     @PostMapping("/money/share")
     public Mono<ResponseEntity<TokenResponseDTO>> createToken(
-            @RequestHeader("X-USER-ID") Integer userId,
+            @RequestHeader("X-USER-ID") long userId,
             @RequestHeader("X-ROOM-ID") String roomId,
             @RequestBody SharingRequestDTO sharingRequestDTO){
         log.debug("received {}", sharingRequestDTO);
